qwen3.8-next: use int64 token offsets in the HC, PLE and PLE-gather kernels
The kernels multiplied an int32 program id by the row width, which wraps
past 2**31 elements: 209,716 tokens for the 4 x 2560 HC/PLE rows and only
~52K tokens for the PLE gather (16 hash heads x 2560 per token), where the
wrapped offset silently returns wrong embeddings. The checkpoint allows
262,144 positions. Promote the ids to int64 before the multiply.

+++ miles_plugins/models/qwen3_8_next/ops/kernel/ple_gather.py
@@ -19,7 +19,8 @@ def _gather_ple_rows_from_pinned(
     row_end,
     BLOCK_D: tl.constexpr,
 ):
-    row_id = tl.program_id(0)
+    # ids hold T * heads rows of embedding_dim each; int32 row offsets wrap past ~52K tokens
+    row_id = tl.program_id(0).to(tl.int64)
     global_idx = tl.load(ids_ptr + row_id)
     in_range = (global_idx >= row_start) & (global_idx < row_end)
     local_idx = tl.where(in_range, global_idx - row_start, 0)
